refactor(web): extract artifact shared atoms (ColoredAvatar, markdown, reactions)
First conversion increment under the green e2e net: pull ColoredAvatar (tokenized,
to components/shared/) + the markdown and reaction helpers (to pages/artifact/lib/)
out of the 3445-line Artifact.tsx. Behavior unchanged — deep artifact spec 4/4 green,
typecheck clean. Seeds the pages/artifact/ module structure for the full split.
